chore: Add "@sanity/client" and "server-only" dependencies
This commit adds the "@sanity/client" and "server-only" dependencies to the "package.json" file. The "@sanity/client" dependency is required for integrating Sanity.io with the project, while the "server-only" dependency is used for server-side operations. These additions enable the implementation of server-side functionality and communication with the Sanity.io API.